@@ -8,32 +8,12 @@
 #include <stdint.h>
 #include <unistd.h>
 
+#include "pstd/include/pstd_string.h"
+
 #include "src/coding.h"
 #include "storage/util.h"
 
 namespace storage {
-
-/* Return the number of digits of 'v' when converted to string in radix 10.
- * See ll2string() for more information. */
-uint32_t Digits10(uint64_t v) {
-  if (v < 10) return 1;
-  if (v < 100) return 2;
-  if (v < 1000) return 3;
-  if (v < 1000000000000UL) {
-    if (v < 100000000UL) {
-      if (v < 1000000) {
-        if (v < 10000) return 4;
-        return 5 + (v >= 100000);
-      }
-      return 7 + (v >= 10000000UL);
-    }
-    if (v < 10000000000UL) {
-      return 9 + (v >= 1000000000UL);
-    }
-    return 11 + (v >= 100000000000UL);
-  }
-  return 12 + Digits10(v / 1000000000000UL);
-}
 
 /* Convert a long long into a string. Returns the number of
  * characters needed to represent the number.
@@ -47,239 +27,19 @@ uint32_t Digits10(uint64_t v) {
  * Modified in order to handle signed integers since the original code was
  * designed for unsigned integers. */
 int Int64ToStr(char* dst, size_t dstlen, int64_t svalue) {
-  static const char digits[201] =
-      "0001020304050607080910111213141516171819"
-      "2021222324252627282930313233343536373839"
-      "4041424344454647484950515253545556575859"
-      "6061626364656667686970717273747576777879"
-      "8081828384858687888990919293949596979899";
-  int negative;
-  uint64_t value;
-
-  /* The main loop works with 64bit unsigned integers for simplicity, so
-   * we convert the number here and remember if it is negative. */
-  if (svalue < 0) {
-    if (svalue != LLONG_MIN) {
-      value = -svalue;
-    } else {
-      value = ((uint64_t)LLONG_MAX) + 1;
-    }
-    negative = 1;
-  } else {
-    value = svalue;
-    negative = 0;
-  }
-
-  /* Check length. */
-  uint32_t const length = Digits10(value) + negative;
-  if (length >= dstlen) return 0;
-
-  /* Null term. */
-  uint32_t next = length;
-  dst[next] = '\0';
-  next--;
-  while (value >= 100) {
-    int const i = (value % 100) * 2;
-    value /= 100;
-    dst[next] = digits[i + 1];
-    dst[next - 1] = digits[i];
-    next -= 2;
-  }
-
-  /* Handle last 1-2 digits. */
-  if (value < 10) {
-    dst[next] = '0' + (uint32_t)value;
-  } else {
-    int i = (uint32_t)value * 2;
-    dst[next] = digits[i + 1];
-    dst[next - 1] = digits[i];
-  }
-
-  /* Add sign. */
-  if (negative) dst[0] = '-';
-  return length;
+  return pstd::ll2string(dst, dstlen, svalue);
 }
 
 /* Convert a string into a long long. Returns 1 if the string could be parsed
  * into a (non-overflowing) long long, 0 otherwise. The value will be set to
  * the parsed value when appropriate. */
 int StrToInt64(const char* s, size_t slen, int64_t* value) {
-  const char* p = s;
-  size_t plen = 0;
-  int negative = 0;
-  uint64_t v;
-
-  if (plen == slen) return 0;
-
-  /* Special case: first and only digit is 0. */
-  if (slen == 1 && p[0] == '0') {
-    if (value != nullptr) *value = 0;
-    return 1;
-  }
-
-  if (p[0] == '-') {
-    negative = 1;
-    p++;
-    plen++;
-
-    /* Abort on only a negative sign. */
-    if (plen == slen) return 0;
-  }
-
-  while (plen < slen && p[0] == '0') {
-    p++;
-    plen++;
-  }
-
-  if (plen == slen) {
-    if (value != nullptr) *value = 0;
-    return 1;
-  }
-
-  /* First digit should be 1-9, otherwise the string should just be 0. */
-  if (p[0] >= '1' && p[0] <= '9') {
-    v = p[0] - '0';
-    p++;
-    plen++;
-  } else if (p[0] == '0' && slen == 1) {
-    *value = 0;
-    return 1;
-  } else {
-    return 0;
-  }
-
-  while (plen < slen && p[0] >= '0' && p[0] <= '9') {
-    if (v > (ULLONG_MAX / 10)) /* Overflow. */
-      return 0;
-    v *= 10;
-
-    if (v > (ULLONG_MAX - (p[0] - '0'))) /* Overflow. */
-      return 0;
-    v += p[0] - '0';
-
-    p++;
-    plen++;
-  }
-
-  /* Return if not all bytes were used. */
-  if (plen < slen) return 0;
-
-  if (negative) {
-    if (v > ((uint64_t)(-(LLONG_MIN + 1)) + 1)) /* Overflow. */
-      return 0;
-    if (value != nullptr) *value = -v;
-  } else {
-    if (v > LLONG_MAX) /* Overflow. */
-      return 0;
-    if (value != nullptr) *value = v;
-  }
-  return 1;
+  return pstd::string2int(s, slen, value);
 }
 
 /* Glob-style pattern matching. */
 int StringMatch(const char* pattern, int pattern_len, const char* str, int string_len, int nocase) {
-  while (pattern_len) {
-    switch (pattern[0]) {
-      case '*':
-        while (pattern[1] == '*') {
-          pattern++;
-          pattern_len--;
-        }
-        if (pattern_len == 1) return 1; /* match */
-        while (string_len) {
-          if (StringMatch(pattern + 1, pattern_len - 1, str, string_len, nocase)) return 1; /* match */
-          str++;
-          string_len--;
-        }
-        return 0; /* no match */
-        break;
-      case '?':
-        if (string_len == 0) return 0; /* no match */
-        str++;
-        string_len--;
-        break;
-      case '[': {
-        int not_flag, match;
-        pattern++;
-        pattern_len--;
-        not_flag = pattern[0] == '^';
-        if (not_flag) {
-          pattern++;
-          pattern_len--;
-        }
-        match = 0;
-        while (1) {
-          if (pattern[0] == '\\') {
-            pattern++;
-            pattern_len--;
-            if (pattern[0] == str[0]) match = 1;
-          } else if (pattern[0] == ']') {
-            break;
-          } else if (pattern_len == 0) {
-            pattern--;
-            pattern_len++;
-            break;
-          } else if (pattern[1] == '-' && pattern_len >= 3) {
-            int start = pattern[0];
-            int end = pattern[2];
-            int c = str[0];
-            if (start > end) {
-              int t = start;
-              start = end;
-              end = t;
-            }
-            if (nocase) {
-              start = tolower(start);
-              end = tolower(end);
-              c = tolower(c);
-            }
-            pattern += 2;
-            pattern_len -= 2;
-            if (c >= start && c <= end) match = 1;
-          } else {
-            if (!nocase) {
-              if (pattern[0] == str[0]) match = 1;
-            } else {
-              if (tolower(static_cast<int>(pattern[0])) == tolower(static_cast<int>(str[0]))) match = 1;
-            }
-          }
-          pattern++;
-          pattern_len--;
-        }
-        if (not_flag) match = !match;
-        if (!match) return 0; /* no match */
-        str++;
-        string_len--;
-        break;
-      }
-      case '\\':
-        if (pattern_len >= 2) {
-          pattern++;
-          pattern_len--;
-        }
-        /* fall through */
-      default:
-        if (!nocase) {
-          if (pattern[0] != str[0]) return 0; /* no match */
-        } else {
-          if (tolower(static_cast<int>(pattern[0])) != tolower(static_cast<int>(str[0]))) return 0; /* no match */
-        }
-        str++;
-        string_len--;
-        break;
-    }
-    pattern++;
-    pattern_len--;
-    if (string_len == 0) {
-      while (*pattern == '*') {
-        pattern++;
-        pattern_len--;
-      }
-      break;
-    }
-  }
-  if (pattern_len == 0 && string_len == 0) return 1;
-  return 0;
+  return pstd::stringmatchlen(pattern, pattern_len, str, string_len, nocase);
 }
 
 int StrToLongDouble(const char* s, size_t slen, long double* ldval) {
